fix build

+++ src/Access/Common/JWKSProvider.cpp
@@ -25,7 +25,7 @@ JWKSType JWKSClient::getJWKS()
     auto now = std::chrono::high_resolution_clock::now();
     auto diff = std::chrono::duration<double>(now - last_request_send).count();
 
-    if (diff < refresh_timeout && cached_jwks.has_value())
+    if (diff < static_cast<double>(refresh_timeout) && cached_jwks.has_value())
         return cached_jwks.value();
 
     Poco::Net::HTTPResponse response;
